Reorder some components and add the two modals. One on the add answer and one on the add question.
