+++ src/components/QnA/QnA.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import axios from 'axios';
+import More from './More.jsx';
 import Container from './styles/Container.styled.js';
-import Search from './comps/Search.jsx';
-import Qlist from './comps/Qlist.jsx';
-import More from './comps/More.jsx';
+import Search from './Search.jsx';
+import Qlist from './Qlist.jsx';
 import TopQA from './styles/TopQA.styled.js';
 import NoQues from './styles/NoQues.styled.js';
 import sampleData from './sampleData';
